Semantic-motor-seeker/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
python-multipart>=0.0.6
sentence-transformers>=2.2.2
qdrant-client>=1.12.0
redis>=5.0.1
pypdf>=3.9.0
python-docx>=1.1.0
openpyxl>=3.1.2
numpy>=1.26.0
pydantic>=2.4.2
httpx>=0.24.1
